6/1/2026 Progress Update
- Schism #59 Created ViewModel tab classes in accordance with an abstract tab class as the base. This helps organize what will be contained in each tab, and also keeps all involved ViewModels from becoming too monolithic, as the HomeViewModel was before.
- Length parameter removed from UI control. This now initializes as 500, but will eventually need to be dynamically configured with respect to the startAddress, as well as the amount of coils/registers made available by the server device (since requesting too many coils/registers makes the MODBUS reading fail).
- ThemesController no longer defines its own Singleton, thus behaving in accordance with my cleaner Dependency Injection control.
- Moved hex string conversion from service config control to WPF config control, since it should be handled further up the chain, so to speak.

+++ Schiism/Schiism.Core/Abstractions/IPC/States/IConfigState.cs
@@ -19,9 +19,10 @@ namespace Schiism.Core.Abstractions.IPC.States
         public string IPAddress { get; }
 
         /// <summary>
-        /// Gets Data Length in accordance with DataSize and StartAddress for min and max allowable value respectively.
+        /// Gets Data Length in accordance with DataSize and StartAddress for min and max allowable values respectively.
+        /// NOTE: Can NOT be controlled via UI.
         /// </summary>
-        public byte DataLength { get; }
+        public ushort DataLength { get; }
 
         /// <summary>
         /// Gets Starting Address.
